refactor get_timeframe out of backtesting class
